implement google.protobuf.Struct de/serialize
`prost_types::Struct` does not implement `Serialize`/`Deserialize`,
which is required by DCS-gRPC to be able to convert between Rust and Lua
types.

It is not possible to implement traits for foreign types. Which is why
this commit adds `#[serde(with = "crate::utils::proto_struct")]` (in
`stubs/build.rs`) to all fields of the type `prost_types::Struct`.

This annotation needs to be manually added for all future fields of this
type.